+++ next.config.ts
@@ -1,0 +1,36 @@
+import { NextConfig } from "next";
+
+const nextConfig: NextConfig = {
+    rewrites: async () => {
+        return [
+            {
+                source: "/api/:path*",
+                destination: process.env.NODE_ENV === 'development' ? "http://127.0.0.1:5000/api/:path*" : "/api/"
+            }
+        ]
+    },
+    async headers() {
+        return [
+            {
+                // matching all API routes
+                // https://vercel.com/guides/how-to-enable-cors
+                source: "/api/:path*",
+                headers: [
+                    { key: "Access-Control-Allow-Credentials", value: "true" },
+                    { key: "Access-Control-Allow-Origin", value: "*" },
+                    {
+                        key: "Access-Control-Allow-Methods",
+                        value: "GET,OPTIONS,PATCH,DELETE,POST,PUT",
+                    },
+                    {
+                        key: "Access-Control-Allow-Headers",
+                        value:
+                            "X-CSRF-Token, X-Requested-With, Accept, Accept-Version, Content-Length, Content-MD5, Content-Type, Date, X-Api-Version",
+                    },
+                ],
+            },
+        ];
+    },
+}
+
+export default nextConfig;