+++ tailwind.config.ts
@@ -1,0 +1,9 @@
+import type { Config } from "tailwindcss";
+
+const config = {
+    content: [
+        "./app/**/*.{js,ts,jsx,tsx}",
+    ],
+} satisfies Config;
+
+export default config;